# mule-shell-model/main-swagger/swagger.yaml
swagger: '2.0'
info:
  title: Mule Shell API
  version: "1.0.0"
# the domain of the service
host: com.stackoverflow
...
schemes:
  - https
# will be prefixed to all paths
basePath: /question
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /letter:
    get:
      operationId: getLetter
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          required: true
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/Letter'
...
definitions:
...
  Letter:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
...